+++ spring-boot-starter/spring-boot-starter-web/src/test/java/com/getyourguide/openapi/validation/integration/ExceptionsNoExceptionHandlerTest.java
@@ -8,7 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.getyourguide.openapi.validation.integration.exception.WithResponseStatusException;
 import com.getyourguide.openapi.validation.integration.exception.WithoutResponseStatusException;
-import com.getyourguide.openapi.validation.integration.openapi.TestViolationLogger;
+import com.getyourguide.openapi.validation.test.openapi.TestViolationLogger;
 import jakarta.servlet.ServletException;
 import java.util.Optional;
 import org.hamcrest.Matchers;
