+++ spring-boot-starter/spring-boot-starter-web/src/test/java/com/getyourguide/openapi/validation/integration/SpringBootTestConfiguration.java
@@ -1,9 +1,12 @@
 package com.getyourguide.openapi.validation.integration;
 
+import com.getyourguide.openapi.validation.api.log.ViolationLogger;
+import com.getyourguide.openapi.validation.test.openapi.TestViolationLogger;
 import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigurationExcludeFilter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.context.TypeExcludeFilter;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
 
@@ -14,4 +17,8 @@ import org.springframework.context.annotation.FilterType;
     @ComponentScan.Filter(type = FilterType.CUSTOM, classes = AutoConfigurationExcludeFilter.class)
 })
 public class SpringBootTestConfiguration {
+    @Bean
+    public ViolationLogger testViolationLogger() {
+        return new TestViolationLogger();
+    }
 }
