+++ test/test-utils/build.gradle
@@ -1,0 +1,16 @@
+import org.springframework.boot.gradle.plugin.SpringBootPlugin
+
+plugins {
+    id 'java-library'
+    alias(libs.plugins.spring.boot) apply false
+}
+
+dependencies {
+    implementation platform(SpringBootPlugin.BOM_COORDINATES)
+
+    compileOnly project(':openapi-validation-api')
+    compileOnly project(':openapi-validation-core')
+
+    // TODO use spotbugs instead and also apply to all modules?
+    implementation(libs.find.bugs)
+}
